requirements.txt
paramiko==2.4
pyyaml==3.13
